+++ MANIFEST.in
@@ -1,0 +1,14 @@
+include README.md
+include LICENSE
+include requirements.txt
+include pyproject.toml
+recursive-include examples *.py *.md
+recursive-include docs *.md *.rst *.txt
+recursive-include tests *.py
+recursive-include chaincraft *.py
+global-exclude *.pyc
+global-exclude *.pyo
+global-exclude *.db
+global-exclude __pycache__
+global-exclude .pytest_cache
+global-exclude *.egg-info
