+++ chaincraft/shared_message.py
