+++ tests/test_chatroom_protocol.py
@@ -10,7 +10,7 @@ from shared_message import SharedMessage
 from examples.chatroom_protocol import ChatroomObject
 
 # Import ECDSA for signing and verifying
-from crypto_primitives.sign import ECDSASignaturePrimitive
+from chaincraft.crypto_primitives.sign import ECDSASignaturePrimitive
 
 
 def get_post_messages(chat_obj, room_name):
